fix grid sizing clamping

+++ src/Core/Components/GridComponent.cs
@@ -2,10 +2,10 @@ using System.Numerics;
 
 namespace Zinc;
 
-public record struct GridComponent(Vector2 GridPivot, Vector2 CellPivot, int CellWidth = 8, int CellHeight = 8, int NumHorizonalCells = 8, int NumVerticalCells = 8) : IComponent
+public record struct GridComponent(Vector2 GridPivot, Vector2 CellPivot, float CellWidth = 8, float CellHeight = 8, int NumHorizonalCells = 8, int NumVerticalCells = 8) : IComponent
 {
-    public int GridWidth => CellWidth * NumHorizonalCells;
-    public int GridHeight => CellHeight * NumVerticalCells;
+    public float GridWidth => CellWidth * NumHorizonalCells;
+    public float GridHeight => CellHeight * NumVerticalCells;
     //square grid runs from 0,0 top left to 1,1 bottom right
     public void GetGridPosition(int index, out float x, out float y)
     {
